requirements.txt
# requirements.txt
kafka-python==2.0.2
elasticsearch==7.17.9
scikit-learn==1.3.2
numpy==1.26.2
paramiko==3.3.1
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1
loguru==0.7.2
